+++ androidApp/build.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 koinCompiler {
     userLogs = true
     debugLogs = true
+    compileSafety = false
 }
 
 android {
