+++ composeApp/src/androidMain/kotlin/com/grappim/taigamobile/di/AndroidAppModule.kt
@@ -53,7 +53,6 @@ import org.koin.core.annotation.Module
 
 @Module(
     includes = [
-        AppModule::class,
         KmpNetworkModule::class,
         DateTimeModule::class,
         DecimalFormatterModule::class,
